Increased scalability and ease of editing
By storing all of the used values for the programm in variables, ease of editing them is increased, as you dont have to go deep into the code anymore and hard code the changes. Also, the integers used for the dividing of the start number have been stored in a list, which makes the programm way more scalable, as new additions dont have to be added via adding new variables. This also makes the programm look more clear. By storing everything inside a function, it is possible to reuse the whole programm by just calling the function again, which also makes makes the programm more scalable.